add UI for user to input responses to game events
Most of this change pertains to inputing chainspaces.
The user clicks on one of the shown spaces to select
that space as a response to the current event.

As for everything else, a select widget is used to list
out various responses that can be selected.

Safe code is put in place to make sure the user has selected
a response before executing the event.