add radiogroup to show various spaces on board ui
A slight rework was implemented.
We now have a SpaceController to handle events involing
spaces in the UI. Clicked on spaces now report to the
controller for further processing.

The controller has 3 modes of operation: Normal, Player, and Spaces.
Normal mode shows no spaces unless the pscore simulator requires
a response in the form of a chainspace. Player mode shows no board
spaces, but adds circles positioned to where the players are on the
board. Spaces mode shows every space on the board.

While spaces shown in Player and Spaces mode are clickable, the
controller ignores requests to process them when in those modes.

Along with those changes, I took the liberty to add [struct].[method]
tags on log messages to make reading easier.